-moved error codes from  ConnectConstants to PersonalIdApiHandler
